Fixes on permissionssRequest's functionality
